feat: updated StudentService to include counting students by sex
Added a new method countBySex to the StudentService interface to count the number of students by sex.
Updated the StudentServiceImpl class to implement the countBySex method.
Changed the searchStudentByName method in the StudentController class to use a query parameter instead of a path parameter.
Added input validation to the assignClassroomToStudent method in the StudentController class to check for invalid ids.
Updated the StudentService interface to include the countBySex method.