feat: update classroom service and controller
Updated Classroom entity to use fetch = FetchType.LAZY for the students field.
Renamed methods in ClassroomService and ClassroomServiceImpl to follow a consistent naming convention:
* getAllClasses -> listAll
* createClass -> create
* updateClass -> update
* deleteClassroom -> delete
* getClassroom -> get
* uploadClassroomProfilePicture -> uploadProfilePicture
* searchClassroomByName -> searchByName
Updated ClassroomController to use the new method names and added @Transactional annotations to the updateClassroom and createClassroom methods.
Updated the ModelMapper configuration to ignore persistent collections.
Added a PropertyMap to the ModelMapper to handle the mapping of Classroom to ClassroomDTO.
Updated the createClassroom method to set the updatedAt field of the Classroom entity.
Updated the deleteClassroom method to use the delete method of the ClassroomService.