feat: introduce GuardianInput DTO and update GuardianController and GuardianServiceImpl
Introduced a new GuardianInput DTO to represent the input data for creating a new guardian.
Updated GuardianController to use the GuardianInput DTO instead of GuardianDTO for creating a new guardian.
Updated GuardianServiceImpl to use persistAndFlush instead of persist to ensure that the guardian is properly saved to the database.
Updated AbsentSchedule to pass an additional parameter false to the createAttendance method.
Updated StudentServiceImpl to use the get method of ClassroomService instead of getClassroom to retrieve a classroom by ID.
Added a @Transactional annotation to the assignStudentToClassroom method in StudentServiceImpl to ensure that the database transaction is properly managed.